Rename repo-package-info to create the start of the new command
